+++ src/compas_viewer/__init__.py
@@ -41,3 +41,4 @@ __all_plugins__ = [
 # Putting imports here to avoid circular imports
 from .viewer import Viewer  # noqa: F401, E402
 from .controller import Controller  # noqa: F401, E402
+from .scene import Tag  # noqa: F401, E402
