+++ src/compas_viewer/actions/info.py
@@ -1,4 +1,4 @@
-from compas_viewer.utilities import gl_info
+from compas_viewer.gl import gl_info
 
 from .action import Action
