--- src/compas_viewer/utilities/__init__.py
@@ -1,3 +1,0 @@
-from .gl import gl_info, make_vertex_buffer, make_index_buffer, update_vertex_buffer, update_index_buffer  # noqa: F401
-from .qt import key_mapper, Timer  # noqa: F401
-from .maps import instance_colors_generator  # noqa: F401
